consolidates all web scrapers
