+++ server/routes.py
@@ -2,11 +2,11 @@ from . import api
 from server.controllers.companies import CompaniesAPI, CompanyAPI
 from server.controllers.jobs import JobsAPI, JobAPI
 from server.controllers.people import PeopleAPI, PersonAPI
-from server.controllers.getjobs import GetJobsAPI
+# from server.controllers.getjobs import GetJobsAPI
 from server.controllers.users import UsersAPI
 from server.controllers.sessions import SessionsAPI
-from server.controllers.people_scraper import PeopleScraperAPI
-from server.controllers.company_scraper import CompanyScraperAPI
+from server.controllers.external_resources import CompanyScraperAPI, PeopleScraperAPI, GetJobsAPI
+# from server.controllers.company_scraper import CompanyScraperAPI
 
 api.add_resource(
     CompanyScraperAPI,
